requirements.txt
# Runtime dependencies (production)
openai>=1.30.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
loguru>=0.7.0
dash>=2.14.0
plotly>=5.18.0
pandas>=2.0.0
requests>=2.31.0
httpx>=0.25.0
tenacity>=8.2.0
